[docker][s]: add apiproxy service with proxy of /auth to auth service.
NOTE: we had a consistent issue of the stack failing to redeploy with
error:

Status 422 (POST
https://cloud.docker.com/api/app/v1/datopian/stack/508fa675-59cc-493a-a843-c6df6afe718b/redeploy/).
Response: {"error": "There is a pending action 'Stack Update' on this stack"}

auth and postgres were working fine but haproxy was down. We had to start haproxy manually. After that we continue to get this message but the stack was up.